Refactor form into more components
